🔨 extract ChartHit into its own file
