+++ packages/@ourworldindata/components/src/styles/util.scss
@@ -26,6 +26,10 @@
     text-align: right;
 }
 
+.list-style-none {
+    list-style: none;
+}
+
 .display-block {
     display: block;
 }
