+++ site/owid.scss
@@ -40,6 +40,7 @@
 @import "./SiteSearchNavigation.scss";
 @import "./DataCatalog.scss";
 @import "./site/search/Autocomplete.scss";
+@import "./site/search/ChartHit.scss";
 @import "./site/search/Search.scss";
 @import "./SiteLogos.scss";
 @import "./SiteNavigation.scss";
