+++ site/search/ChartHit.scss
@@ -1,0 +1,91 @@
+.chart-hit {
+    &:hover {
+        .chart-hit-highlight {
+            text-decoration: underline;
+        }
+        .chart-hit-img-container img {
+            box-shadow: none;
+        }
+    }
+}
+
+.chart-hit-title-container {
+    line-height: 20px;
+
+    .chart-hit-highlight {
+        color: $blue-90;
+    }
+
+    .chart-hit-variant {
+        color: $blue-60;
+        font-size: 0.9em;
+    }
+}
+
+@keyframes chartErrorFadeIn {
+    from {
+        opacity: 0;
+    }
+    to {
+        opacity: 1;
+    }
+}
+
+.chart-hit-img-container {
+    margin-bottom: 8px;
+    aspect-ratio: calc($grapher-thumbnail-width / $grapher-thumbnail-height);
+    width: 100%;
+    background-color: $gray-10;
+
+    img {
+        display: block; // remove the space below the image
+        transition: box-shadow 0.1s;
+
+        &.loaded {
+            box-shadow:
+                0px 0px 0px 0px rgba(49, 37, 2, 0.03),
+                0px 6px 13px 0px rgba(49, 37, 2, 0.03),
+                0px 93px 37px 0px rgba(49, 37, 2, 0.01),
+                0px 145px 41px 0px rgba(49, 37, 2, 0);
+        }
+
+        &.error {
+            display: none;
+        }
+    }
+
+    .chart-hit-img-error {
+        display: flex;
+        height: 100%;
+        flex-direction: column;
+        align-items: center;
+        justify-content: center;
+
+        color: $grey-text-color;
+        animation: chartErrorFadeIn 0.3s;
+
+        gap: 0.5em;
+
+        svg {
+            font-size: 1.75em;
+        }
+    }
+}
+
+.chart-hit-entities {
+    list-style: none;
+    font-size: 0.8em;
+
+    li {
+        display: inline;
+        color: $blue-50;
+
+        &::after {
+            content: ", ";
+        }
+
+        &:last-child::after {
+            content: "";
+        }
+    }
+}
